backend/requirements.txt
# Core dependencies
fabric-sdk-py==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2

# Fix protobuf version
protobuf==3.20.3
grpcio==1.44.0

# Cryptography
cryptography==41.0.3

# Storage
ipfshttpclient==0.8.0 